+++ src/AppserverIo/Appserver/PersistenceContainer/TimedObjectInvoker.php
@@ -32,7 +32,7 @@ use AppserverIo\Psr\EnterpriseBeans\TimedObjectInterface;
 use AppserverIo\Psr\EnterpriseBeans\TimedObjectInvokerInterface;
 use AppserverIo\Psr\EnterpriseBeans\Annotations\Timeout;
 use AppserverIo\Psr\EnterpriseBeans\Annotations\Schedule;
-use AppserverIo\Psr\PersistenceContainerProtocol\BeanContext;
+use AppserverIo\Psr\EnterpriseBeans\BeanContext;
 use AppserverIo\Psr\Application\ApplicationInterface;
 
 /**