+++ CHANGELOG.md
@@ -2,12 +2,11 @@
 
 ## Bugfixes
 
-* None
+* Remove unnecessary interfaces SenderInterface + ReceiverInterface
 
 ## Features
 
-* Move interfaces ResourceLocator, ServiceContext, ServiceExecutor, ServiceProvider, ServiceResourceLocator + TimerServiceContext from appserver-io/appserver package
-* Move exceptions InvalidBeanTypeException + ServiceAlreadyRegisteredException from appserver-io/appserver package
+* Refactoring, move interfaces of Persistence-Container + Message-Queue to separate packages
 
 # Version 1.0.0-beta4